Refactor course creation functionality and introduce new hooks
- Replaced the legacy `postNewCourse` function with a new `createNewCourse` function for improved clarity and maintainability.
- Introduced `useCreateNewCourse` hook to handle new course creation with toast notifications and query invalidation.
- Updated `NewCourseForm` to utilize the new `useCreateNewCourse` hook, streamlining the form submission process.
- Removed the outdated `useCreateCourseCopy` hook and integrated its functionality into the new `useCreateCourse` hook for better organization.

These changes enhance the course creation process, improve code structure, and provide a more intuitive user experience.
